+++ lib/aliyun_oss/bucket/lifecycle.ex
@@ -10,6 +10,64 @@ defmodule Aliyun.Oss.Bucket.Lifecycle do
   @type error() :: %Error{body: String.t(), status_code: integer(), parsed_details: map()} | atom()
 
   @doc """
+  PutBucketLifecycle接口用于设置存储空间（Bucket）的生命周期规则。生命周期规则开启后，OSS将按照配置规则指定的时间，自动转换与规则相匹配的文件（Object）的存储类型或将其删除。
+
+  ## Examples
+
+      iex> config = %{
+        "LifecycleConfiguration" => %{
+          "Rule" => %{
+            "AbortMultipartUpload" => %{"Days" => "1"},
+            "Expiration" => %{"Days" => "1"},
+            "ID" => "delete objects and parts after one day",
+            "Prefix" => "logs/",
+            "Status" => "Enabled"
+          }
+        }
+      }
+      iex> Aliyun.Oss.Bucket.Lifecycle.put("some-bucket", config)
+      {:ok, %Aliyun.Oss.Client.Response{
+        data: "",
+        headers: [
+          {"Date", "Wed, 05 Dec 2018 02:34:57 GMT"},
+          ...
+        ]
+      }}
+      iex> config = ~S[
+      <?xml version="1.0" encoding="UTF-8"?>
+      <LifecycleConfiguration>
+        <Rule>
+          <ID>delete objects and parts after one day</ID>
+          <Prefix>logs/</Prefix>
+          <Status>Enabled</Status>
+          <Expiration>
+            <Days>1</Days>
+          </Expiration>
+          <AbortMultipartUpload>
+            <Days>1</Days>
+          </AbortMultipartUpload>
+        </Rule>
+      </LifecycleConfiguration>
+      ]
+      iex> Aliyun.Oss.Bucket.Lifecycle.put("some-bucket", config)
+      {:ok, %Aliyun.Oss.Client.Response{
+        data: "",
+        headers: [
+          {"Date", "Wed, 05 Dec 2018 02:34:57 GMT"},
+          ...
+        ]
+      }}
+  """
+  @spec put(String.t(), String.t() | map()) :: {:error, error()} | {:ok, Response.t()}
+  def put(bucket, %{} = config) do
+    put(bucket, MapToXml.from_map(config))
+  end
+
+  def put(bucket, config) do
+    put_bucket(bucket, %{}, %{"lifecycle" => nil}, config)
+  end
+
+  @doc """
   GetBucketLifecycle 用于查看Bucket的Lifecycle配置。
 
   ## Examples