+++ CHANGELOG.md
@@ -5,6 +5,7 @@
   - New APIs:
     - GetBucketV2 (ListObjectsV2): `Aliyun.Oss.Bucket.list_objects`
     - GetObjectMeta: `Aliyun.Oss.Object.get_object_meta`
+    - PutBucketLifecycle: `Aliyun.Oss.Lifecycle.put`
   - **Breaking changes**:
     - Use separated modules to scope SubResources of Bucket/Object, for example:
       - `Aliyun.Oss.Bucket.get_acl` becomes `Aliyun.Oss.Bucket.ACL.get`